First commit
First commit to this fork. Added retrospective.py which uses GFS archive data for flight predictions in the past. Corrected the coordinate order in get_gfs.py to match the example. Updated get_gfs.py with the new urls for the 1 and 0.5 degree 'main' forecasts. Added 'simple_ascent_rate' to pyb_traj.py to override the ascent speed calculation if required (i.e. ignore the balloon volume and drag coefficient).